refactor: added docs
