+++ crates/proof-of-sql/src/arrow/base_commitment_arrow/mod.rs
@@ -1,1 +1,2 @@
-pub mod table_commitment;
+/// This module contains errors that can occur while manipulating a [`TableCommitment`].
+pub mod table_commitment_errors;
