+++ crates/proof-of-sql/src/arrow/base_commitment_arrow/table_commitment_errors.rs
