+++ crates/proof-of-sql/src/arrow/base_database_arrow/column.rs
