+++ crates/proof-of-sql/src/arrow/mod.rs
@@ -1,3 +1,10 @@
+//! This module contains the modules and functions which uses the crate Arrow
+
+/// This module contains types for creation and utilization of cryptographic commitments to proof-of-sql data that uses Arrow
 pub mod base_commitment_arrow;
+
+/// This module contains modules with database related functionality that uses Arrow
 pub mod base_database_arrow;
+
+/// This module contains modules related to sql proof that utilizes Arrow
 pub mod sql_proof_arrow;
