+++ crates/proof-of-sql/src/arrow/sql_proof_arrow/final_round_builder_test.rs
@@ -1,4 +1,3 @@
-
 #[cfg(feature = "arrow")]
 #[test]
 fn we_can_form_the_provable_query_result() {
