+++ crates/proof-of-sql/src/arrow/sql_proof_arrow/mod.rs
@@ -1,4 +1,5 @@
+/// Module for query results
 pub mod query_result;
-pub mod final_round_builder_test;
+mod final_round_builder_test;
 #[cfg(all(test, feature = "arrow"))]
 mod provable_query_result_test;
